requirements.txt
streamlit==1.33.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==9.5.0
